requirements.txt
boto3==1.20.10
python-binance==1.0.15
twilio==7.4.0
